requirements.txt
# Core LangChain + LangGraph (compatible versions)
langchain==0.2.14
langchain-core==0.2.36
langchain-community==0.2.12
langgraph==0.0.56

# Vector Database
chromadb==0.5.5

# Embeddings / ML
sentence-transformers==3.0.1
faiss-cpu==1.8.0
scikit-learn==1.5.1

# LLM Client (Stable combination)
openai==1.33.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
PyYAML==6.0.2

# Optional (nice to have)
tqdm==4.66.5
seaborn==0.13.2
